+++ crates/telio-traversal/src/endpoint_providers/stun.rs
@@ -1938,6 +1938,7 @@ mod tests {
             async fn get_interface(&self) -> Result<Interface, Error>;
             async fn get_adapter_luid(&self) -> Result<u64, Error>;
             async fn wait_for_listen_port(&self, d: Duration) -> Result<u16, Error>;
+            async fn wait_for_proxy_listen_port(&self, d: Duration) -> Result<u16, Error>;
             async fn get_wg_socket(&self, ipv6: bool) -> Result<Option<i32>, Error>;
             async fn set_secret_key(&self, key: SecretKey) -> Result<(), Error>;
             async fn set_fwmark(&self, fwmark: u32) -> Result<(), Error>;